run 2to3 and change submodule repository.

+++ dense_correspondence/loss_functions/pixelwise_contrastive_loss.py
@@ -292,7 +292,7 @@ class PixelwiseContrastiveLoss(object):
         non_match_loss_vec, num_hard_negatives, _, _ = PCL.non_match_descriptor_loss(image_a_pred, image_b_pred, non_matches_a,
                                                                  non_matches_b, M=M_descriptor, invert=invert)
 
-        num_non_matches = long(non_match_loss_vec.size()[0])
+        num_non_matches = int(non_match_loss_vec.size()[0])
 
 
         non_match_loss = non_match_loss_vec.sum()
